+++ src/controller_rpc.c
@@ -937,7 +937,7 @@ xml_add_op(cxobj *x,
  *   5) Modify tree with xmldb_put
  * @param[in]  h    Clixon handle
  * @param[in]  db   Database
- * @param[in]  cvv  Vector of services
+ * @param[in]  cvv  Vector of services, if empty then all
  * @retval     0    OK
  * @retval    -1    Error
  * @note Differentiate between reading created from running while deleting from action-db
@@ -1265,7 +1265,7 @@ controller_commit_actions(clixon_handle           h,
                           actions_type            actions,
                           transaction_data_t     *td,
                           char                   *service_instance,
-			  int                    diff
+			  int                     diff
                           )
 {
     int     retval = -1;
@@ -1283,7 +1283,7 @@ controller_commit_actions(clixon_handle           h,
      */
     if (controller_actions_diff(h, ct, td, &services, cvv) < 0)
         goto done;
-    if (actions == AT_FORCE){
+    if (actions == AT_FORCE || actions == AT_DELETE){
         cvec_reset(cvv);
         if (service_instance)
             cvec_add_string(cvv, service_instance, NULL);
@@ -1297,12 +1297,19 @@ controller_commit_actions(clixon_handle           h,
 #endif
     if (xmldb_copy(h, "candidate", "actions") < 0)
         goto done;
-    if (services &&
-        (actions == AT_FORCE || cvec_len(cvv) > 0)){
+    if (services && actions == AT_DELETE){
+        /* Delete service, do not activate/notify actions, just push deletes to devices
+           Strip service data in device config */
+        if (strip_service_data_from_device_config(h, "actions", cvv) < 0)
+            goto done;
+        if (commit_push_after_actions(h, ct) < 0)
+            goto done;
+    }
+    else if (services && (actions == AT_FORCE || cvec_len(cvv) > 0)){
         /* IF Services exist AND
-           either service changes or forced,
-           THEN notify services
-        */
+         * either service changes or forced,
+         * THEN notify services
+         */
         if (services_commit_notify(h, ct, cvv, diff) < 0)
             goto done;
         /* Strip service data in device config for services that changed */
@@ -1758,6 +1765,7 @@ rpc_controller_commit(clixon_handle h,
         break;
     case AT_CHANGE:
     case AT_FORCE:
+    case AT_DELETE:
         if (strcmp(ct->ct_sourcedb, "candidate") != 0){
             if (netconf_operation_failed(cbret, "application", "Only candidates db supported if actions")< 0)
                 goto done;
@@ -1765,10 +1773,8 @@ rpc_controller_commit(clixon_handle h,
                 goto done;
             goto ok;
         }
-
 	if (pusht == PT_NONE)
 	    diff = 1;
-
         /* Compute diff of candidate, copy to actions, trigger notify */
         if (controller_commit_actions(h, ct, actions, td, service_instance, diff) < 0)
             goto done;