+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ### New features
 
+* New: [RESTCONF service delete does not work](https://github.com/clicon/clixon-controller/issues/199)
+   * Add delete service by extending rpc controller-commit with DELETE
 * New: [NACM for Clixon Controller](https://github.com/clicon/clixon-controller/issues/189)
   * Support for services
   * Testcases for controller nacm and nacm+restconf